refactor: reorganize stories into subfolder and clean up
- move DiagramCanvas stories into stories/ subfolder (theme, components, mock, DiagramCanvas.stories)
- replace argTypes markdown tables with Storybook native type/description format
- remove console.log from handleEdgesChange
- add .github/pull_request_template.md
