+++ .github/pull_request_template.md
@@ -1,0 +1,42 @@
+## Link al ticket
+
+- [JMV-XXXX](https://janiscommerce.atlassian.net/browse/JMV-XXXX)
+
+## Descripción del requerimiento
+
+-
+
+## Criterios de aprobación
+
+-
+
+## Descripción de la solución
+
+-
+
+## ¿Cómo se puede probar?
+
+### Links de prueba:
+
+| Caso a probar | Resultado esperado | Resultado obtenido | Observaciones |
+| ------------- | ------------------ | ------------------ | ------------- |
+|               |                    | ✅ o ❌            |               |
+
+## Evidencias, pruebas de cómo funciona
+
+-
+
+## Link a la documentación
+
+-
+
+## Datos extra a tener en cuenta
+
+-
+
+## CHANGELOG (English only):
+
+```javascript
+### Added
+-
+```
