+++ src/components/DiagramCanvas/stories/theme.js
@@ -1,0 +1,4 @@
+export const PRIMARY = '#2979ff';
+export const PRIMARY_SOFT = '#e8f0ff';
+export const SECONDARY_DEEP = '#b86a2b';
+export const SECONDARY_SOFT = '#ffeddd';
